[Releng] Check size of signed.zip (MacOS)

+++ releng/org.eclipse.oomph.releng/hudson/repackage.sh
@@ -65,6 +65,16 @@ for f in *.zip; do
       zip -r -q unsigned.zip "Eclipse Installer.app"
       rm -rf "Eclipse Installer.app"
       curl -o signed.zip -F file=@unsigned.zip http://build.eclipse.org:31338/macsign.php
+      
+      actualSize=$(wc -c signed.zip | cut -f 1 -d ' ')
+      if [ $actualSize -lt 40000000 ]; then
+        echo "signed.zip is just $actualSize bytes large!"
+        echo ""
+        cat signed.zip
+        echo ""
+        exit 1
+      fi
+      
       unzip -qq signed.zip
       rm -f unsigned.zip signed.zip
     fi
@@ -80,7 +90,7 @@ for f in *.zip; do
       curl -o signed.exe -F filedata=@eclipse-inst.exe http://build.eclipse.org:31338/winsign.php
       mv signed.exe eclipse-inst.exe
 
-      actualSize=$(wc -c "eclipse-inst.exe" | cut -f 1 -d ' ')
+      actualSize=$(wc -c eclipse-inst.exe | cut -f 1 -d ' ')
       if [ $actualSize -lt 200000 ]; then
         echo "eclipse-inst.exe is just $actualSize bytes large!"
         echo ""
@@ -120,7 +130,7 @@ for f in *.zip; do
       if [ $actualSize -lt 40000000 ]; then
         echo "$PRODUCTS/$extractor is just $actualSize bytes large!"
         echo ""
-        cat $PRODUCTS/$extractor
+        cat "$PRODUCTS/$extractor"
         echo ""
         exit 1
       fi
